incorporate CV_photo analysis
fix some minor issue in that analysis and several broader updates,
including allowing files in an ANA to have only 1 semicolon
(file_type;sample_no). Also, in visualization, accounting for the fact
that filed['keys'] may not exist. Also, if any fom has 'NaN' read that
column as float32, which means that an integer FOM will be either int or
float in the visulaizer depending on whether there is a NaN. This isn't
great but probably won't cause issues.