Change run comment and run_use
eliminate new feature for run comment because near impossible to get it
to not apply to every run since the exp gets regenerated on every "add".

similarly the user foms not saved for previous "add" so have the run_use
be, e.g. data__val0,val1,val2  where the user-entered string values make
the run_use unique and therefore separated into different run__ blocks.
a consequence is that a given file could be added to the .exp as data
and data__<fom>.