ana and visualization of 4 element subsets of up to 8 element print
for 6 element multi-pH OER study, define 4 element subsets of a 6
element print that can be plotted as a quaternary space using labels
from database.
platemap_comp4plot_keylist is the ana__ key, and in visualizaer user can
only manually enter 4 elements but the code will handle database sets of
elements, presumably only up to 8 since that is how many channels are in
the platemaps. for >4  element labels, permuations are done with respect
to indeces of the 4-element space after partitioning. Not sure what
happens if try to plat from multiple ana__ that have different 4-element
selections.
also a new plotting feature that uses the "-1" code in the comp plot
tabs to represent "plot everything"