Add option to not save .dat in CreateExp, Cosmetic changes to Visualize
now checkbox in CreateExp that skips saving .dat but even in this case
all raw data files are opened to ensure they exist and to get header
lines and other file attrs.
change uvis batch to fill in "uvis" user and type.
use MAXNUMRAWDATAHEADERBYTES_FORWHENNOTSAVINGBINARY to avoid reading
full files when not saving binary.
In Visualize, display ellabels and first draft of modifying the summary
string to include elemtent subsets.