changed ECHEPHOTO_checkcompletedanalysis_inter_filedlist
this finds files applicable for subsequent analysis, like Pmax and
Iphoto. This is subsequent analysis on an existing ana__ so there must
be only 1 matching ana__ but now it discriminates by technique so if
Iphoto calculated on a technique once then that's ok but if more than
once a message will be printed.
also introduces gui_mode_bool for each anlaysis class to allow analyses
to use dialogs if necessary but allow batch scripts to turn these off