Open exp or ana from .info or search, and save selectsamples in visualizer
OpenFromInfoForm is new and several chanegs to allow flixebility for
folder, .zip, .zip+fn entry in importexp and importana in CalcFOM and
Visualize
In CalcFOM and Visualizer, .exp or .ana can be opened by checking the
desired one from the .info file or by searching all EXP or ANA using a
search string, e.g. the time stamp
Also, in visualizer the select samples broswer text can be saved with
file included in the .ana. The browser has its own checkbox in the "Save
figs to ANA" menu.