fix multi-element ink and start .ana composition handling
multi-element ink mostly vest in Visualizer.
Most changes related to adding aux exp/ana to active .ana and analysis
post process class Analysis__FOM_Merge_Aux_Ana.
This required introduction of aux_ana_dlist in calcfom and in each
analysis class, the passing of calcFOMDialogclass as an instance to that
master class to access it. None of that has been tested