check exp when saving ana
if saving ana in e.g. eche or uvis folder then make sure the exp is in
analogous folder  and isn't .run, because in those cases the .ana will
be set to go to J but will be missing it's exp. User has the option to
cancel the ana save or copy the exp to the appropriate folder with
.done.  If an aux exp or ana is in temp or is .run then ana cannot be
saved