Visualize using FOM compositions
in extractplotdinfo, the user-specified string for searching the FOMs is
used in combination with the element labels (FOM name must start with el
label)  to get up to 4 FOMs (0 is used of no key found), which are
normalized to dcreate the compositions in self.fomplotd, which are used
for plotting and other things. If multiple keys meet criteria, the first
one foudn is used, and in general the availability of keys and which
keys are used may vary from samples to sample. so some samples may have
comps from FOMs but others without the composition keys obtain the
default platemap comps