screening_print_id and platemap4keys_default
look up element list by the screening_print_id, including if it is in a
parent .info.

For visualizing, introduce platemap4keys_default and expand the meaning
of platemap4keys to include any key that is in all the platemap samples,
which inldues elements with calculated concentrations from multielement
inks. For that scenario, still need to choose a 4 element projection for
compsition plots and let the user select 1. exp or ana needs to be
reopened to select another.