Analysis__XRFS_EDAX
this analysis function should perform same actions as the database auto
generated ana__1 for XRFS, with the additional functinailty that
multiple run__ in the exp will be combined in the same ana__1 fom .csv
with the xrfs transitions being the union of all source xrfs runs. The
.CPS .WtPerc .AtPerc appendings to the EDAX column headings can also be
customized in analysis parameters.