first complete draft of Analysis__Process_B_vs_A_ByRun
compares B to A sample by sample where B and A can correspond to rows
from the csv with different (or the same) runint(s)
'select_ana': 'ana__1' is an ana which should have 1 fom csv
'fom_keys_B':'ALL', is default which will add all fom keys. can also be
,-delimited lsit of search strings, as in other process functions
'fom_keys_A':'ALL', better end up with same length as B
'runints_B':'2' consider all rows in the csv with runint in this list
(e.g. 2,3,4) to be B
'runints_A':'1', if this is a subset of B then the comparision is from
the same rows of the csv so if the fom_keys are the same then you'll get
1's. That is make, runints the same for A and B if you want to compare
foms for the same csv rows. If there are different then the intersection
of sample_no within A and B runints will end up int he resulting file.
DOES NOT COMPARE plate_id - assumes map_id are the same such that
pairing by sample_no is meaningful.
'method':'B_over_A', B_over_A for ratio B:A and B_minus_A for B-A are
initially supported
'relative_key_append':'_RelRatio' this has default values but can make
it custom. The results are saved using the sample,run,plate of B and the
keys of B foms with this addition