"plot selected" in visualizer
can select or enter sample_no list and then "plot selecT" button will
use only the browser samples in the plot. click a "standard plot" to
remove this filter.

made _PlotAnaForm_tryLayouts in an attempt to make an effective
sizepolicy but didn't work so this file saved as "documentation" of the
effort.