new analysis, calcfom class functions, ssrl import v3
several things with ssrl import to enable .udi to be written without
compositions if roi comps not available in .h5. Previously ana said the
file was there but it wasn't

ssrl_import_v3.py is now the corrected importer and _xrfs_interp is in
development but not working.

Analysis__FOM_Interp_Merge_Aux_Ana complete draft but not tested

to enable non-dialog classes to have access to running and managing
anadict the same way calcfom does,
Analysis_Master has the master functions
calcfom_analyzedata_calcfomdialogclass, gethighestanak