fix bug in merge, new fcns_io functions, new batch program
. merge wasn't working when using sample_no union due to plate_id
tracking. no change in function of the analysis
. new utility functions for converting dates from .info, and change dict
reader to have fewer assumptions on files_ nested dicts because image
has, e.g. fome_files: ''
. batch program explore_info_files_v1_metrprintrecords.py looks through
all the info and finds all analysis of a given type (imag) and fins the
most recent print and anneal preceding the run__1 in the exp referenced
in the ana.