fix Marge analysis via UI
getapplicablefilenames called when analysis function selected and
assumes processnewmarams is called for process analyses, but in
Analysis_FOM_Merge_Aux_Ana this was only done for 'custom', probably to
save overhead from repeated recalculation. But this caused the user
interface version to not work. A separate issue is that the aux ana path
was split and I don't think it should be (line 41)