guimode  changes to CalcFOMApp
don't make all the connect button links if guimode=False (check that
batch analysis could be done with guimode=False using the ecms batch
script). The motivation was because spyder was having a problem with
intiializing the form.

add section to ecms to enable somewhat indepednet running for debug
purposes